// resto-back/src/auth/entities/permissions.enum.ts
export enum Permissions {
    MANAGE_STAFF = 'MANAGE_STAFF',           
    MANAGE_MENU = 'MANAGE_MENU',             
    MANAGE_INVENTORY = 'MANAGE_INVENTORY',   
    VIEW_REPORTS = 'VIEW_REPORTS',           
    VIEW_MENU = 'VIEW_MENU',                 
    TAKE_ORDERS = 'TAKE_ORDERS',             
    UPDATE_INVENTORY = 'UPDATE_INVENTORY'    
  }